% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{all}
\alias{all}
\title{Smoking score calculated using MS, SSc and MLM methods.}
\usage{
all(dataset, dataset_QN, dataset_ILM, dataset_SQN, samplesheet, ref.Elliott,
  ref.Zhang, ref.CS, ref.FS, ref.NS)
}
\arguments{
\item{dataset}{A data matrix of normalised methylation values in beta scale, with rows labelling the CpG probe ids and columns labelling the sample names.
In MLM missing methylation values are imputed as 0.5. In MS and SSc missing CpGs are excluded from the calculation.
eg: dummyBetaData}

\item{samplesheet}{A dataframe with samplenames as rownames. Must contain gender information as  "sex" column. sex column should be marked as 1 and 2
representing male and female respectively. Missing "sex" information is imputed as 0.
eg: dummySamplesheet}

\item{ref.Elliott}{This is used in "SSc" and "all" methods. It is a dataframe with 187 CpGs and the effect sizes used to calculate the smoking score
proposed by Elliott et al PMID: 24485148.}

\item{ref.Zhang}{This is used in "MS" and "all" methods. It is a dataframe with 4 CpGs and weights used to calculate methylation score proposed by Zhang et al PMID: 26826776.}

\item{ref.CS}{This is used in "SSt" and "all" methods. It is a dataframe with 121 CpGs selected by Multinomial LASSO approach, with log odd values for "CURRENT SMOKER" class.
In addition intercept term and sex coefficient terms are also provided by Multinomial LASSO approach.}

\item{ref.FS}{This is used in "SSt" and "all" methods. It is a dataframe with 121 CpGs selected by Multinomial LASSO approach, with log odd values for "FORMER SMOKER" class.
In addition intercept term and sex coefficient terms are also provided by Multinomial LASSO approach.}

\item{ref.NS}{This is used in "SSt" and "all" methods. It is a dataframe with 121 CpGs selected by Multinomial LASSO approach, with log odd values for "NEVER SMOKER" class.
In addition intercept term and sex coefficient terms are also provided by Multinomial LASSO approach.}
}
\value{
Returns a result object with the methylaion score generated by MS, smoking score generated by SSc and predicted smoking probabilities and smoking status
labels generated from MLM method.
}
\description{
Estimates smoking score using three methods: MS, SSc and MLM.
}
\examples{
data(dummyBetaData)
samplesheet <- read.csv( system.file("extdata", "samplesheet_GSE42861.csv", package= "EpiSmokEr"), header=TRUE, sep=",")
result <- epismoker(dataset = dummyBetaData, samplesheet = samplesheet, ref.Elliott =  Illig_data, ref.Zhang = Zhangetal_cpgs,
                    ref.CS = CS_final_coefs, ref.FS = FS_final_coefs, ref.NS = NS_final_coefs, method = "all")
## result contains smoking score calculated from all the three methods ( "SSc", "MS", "SSt").
}
