% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM}
\alias{EM}
\title{Smoking score calculation using Elliott Method.}
\usage{
EM(dataset, ref.Elliott = Illig_data)
}
\arguments{
\item{dataset}{A data matrix of normalised methylation values in beta scale, with rows labelling the CpG probe ids and columns labelling the sample names.
Missing CpGs are excluded from the smoking score calculation.
eg: dummyBetaData}

\item{ref.Elliott}{This is used in "EM" and "all" methods. It is a dataframe with 187 CpGs and the effect sizes used to calculate the smoking score
proposed by Elliott et al PMID: 24485148.}
}
\value{
Returns a result object with smoking score generated by following approach outlined in Elliott et al PMID: 24485148.
}
\description{
Estimates smoking score using Elliott Method.
}
\examples{
data(dummyBetaData)
result <- epismoker(dataset = dummyBetaData, method = "EM")
## result contains smoking score calculated using EM.
}
