% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Illig_data.R
\docType{data}
\name{Illig_data}
\alias{Illig_data}
\title{Illig_data}
\format{A dataframe with 187 rows and 9 columns.}
\usage{
data(Illig_data)
}
\description{
Illig_data contains 187 CpGs required to calcualte smoking score based on Elliot et al approach
}
\examples{
data(Illig_data)
head(Illig_data)
}
\references{
Elliott HR, Tillin T, McArdle WL, et al. Differences in smoking associated DNA methylation patterns in
South Asians and Europeans. Clinical Epigenetics. 2014;6(1):4.PMID: 24485148.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3915234/}{PMC}
}
\keyword{datasets}
