% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Reads idat files}
\usage{
loadData(idatPath)
}
\arguments{
\item{idatPath}{Requires a path to the folder with the idat files and sample sheet included.}
}
\value{
dataset
A RGSet object
}
\description{
loadData function reads idat files and saves as an RGSet object
}
\examples{
rawdata <- loadData(idatPath = system.file("extdata", package = "EpiSmokEr"))
## rawdata is a RGSet object with raw intensity values.

}
