% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS_final_coefs.R
\docType{data}
\name{NS_final_coefs}
\alias{NS_final_coefs}
\title{NS_final_coefs}
\format{A vector of 123 elements.}
\usage{
data(NS_final_coefs)
}
\description{
NS_final_coefs contains 121 CpGs  and their corresponding weights required to calcualte smoking probabilities and predict smoking status
based on Multinomial LASSO method. In addition to CpGs this object also contains intercept and sex terms.
}
\examples{
data(NS_final_coefs)
head(NS_final_coefs)
}
\keyword{datasets}
