% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSt.R
\name{SSt}
\alias{SSt}
\title{Estimates smoking probabilities and smoking status (SSt) using MLM method.}
\usage{
SSt(dataset, samplesheet, ref.CS = CS_final_coefs,
  ref.FS = FS_final_coefs, ref.NS = NS_final_coefs)
}
\arguments{
\item{dataset}{A data matrix of normalised methylation values in beta scale, with rows labelling the CpG probe ids and columns labelling the sample names.
Missing methylation values are imputed as 0.5.
eg: dummyBetaData}

\item{samplesheet}{A dataframe with samplenames as rownames. Must contain gender information as  "sex" column. sex column should be marked as 1 and 2
representing male and female respectively. Missing "sex" information is imputed as 0.
eg: dummySamplesheet}

\item{ref.CS}{This is used in "SSt" and "all" methods. It is a dataframe with 121 CpGs selected by Multinomial LASSO approach, with log odd values for "CURRENT SMOKER" class.
In addition intercept term and sex coefficient terms are also provided by Multinomial LASSO approach.}

\item{ref.FS}{This is used in "SSt" and "all" methods. It is a dataframe with 121 CpGs selected by Multinomial LASSO approach, with log odd values for "FORMER SMOKER" class.
In addition intercept term and sex coefficient terms are also provided by Multinomial LASSO approach.}

\item{ref.NS}{This is used in "SSt" and "all" methods. It is a dataframe with 121 CpGs selected by Multinomial LASSO approach, with log odd values for "NEVER SMOKER" class.
In addition intercept term and sex coefficient terms are also provided by Multinomial LASSO approach.}
}
\value{
Returns a result object with predicted smoking probabilities and smoking status labels generated from MLM method.

data(dummyBetaData)
samplesheet <- read.csv( system.file("extdata", "samplesheet_GSE42861.csv", package= "EpiSmokEr"), header=TRUE, sep=",")
result <- epismoker(dataset = dummyBetaData, samplesheet = samplesheet, ref.CS = CS_final_coefs, ref.FS = FS_final_coefs, ref.NS = NS_final_coefs, method = "SSt")
## result contains predicted smoking probabilities and smoking status labels generated from MLM method.
}
\description{
Estimates smoking probabilities and smoking status (SSt) using MLM method.
}
