% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhangetal_cpgs.R
\docType{data}
\name{Zhangetal_cpgs}
\alias{Zhangetal_cpgs}
\title{Zhangetal_cpgs}
\format{A vector with four elements.}
\usage{
data(Zhangetal_cpgs)
}
\description{
Zhangetal_cpgs contains 4 CpGs  and their corresponding weights required to calcualte methylation score based on Zhang et al approach
}
\examples{
data(Zhangetal_cpgs)
head(Zhangetal_cpgs)
}
\references{
Yan Zhang, Ines Florath, Kai-Uwe Saum, Hermann Brenner.
Self-reported smoking, serum cotinine, and blood DNA methylation. Environmental Research, Volume 146, 2016.
\href{https://www.ncbi.nlm.nih.gov/pubmed/26826776}{Pubmed}
}
\keyword{datasets}
