% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZM.R
\name{ZM}
\alias{ZM}
\title{Methylation score calculation using Zhang Method.}
\usage{
ZM(dataset, ref.Zhang = Zhangetal_cpgs)
}
\arguments{
\item{dataset}{A data matrix of normalised methylation values in beta scale, with rows labelling the CpG probe ids and columns labelling the sample names.
Missing CpGs are excluded from the methylation score calculation.
eg: dummyBetaData}

\item{ref.Zhang}{This is used in "ZM" and "all" methods. It is a dataframe with 4 CpGs and weights used to calculate methylation score proposed by Zhang et al PMID: 26826776.}
}
\value{
Returns a result object with smoking score generated by following approach outlined in Zhang et al PMID: 26826776.
}
\description{
Estimates methylation score using Zhang Method.
}
\examples{
data(dummyBetaData)
result <- epismoker(dataset = dummyBetaData, ref.Zhang = Zhangetal_cpgs, method = "ZM")
## result contains methylation score calculated from ZM.
}
